+++ lib/shared/src/configuration.ts
@@ -30,7 +30,7 @@ export interface Configuration {
     experimentalChatPanel: boolean
     experimentalChatPredictions: boolean
     experimentalSearchPanel: boolean
-    experimentalCommandLenses: boolean
+    commandCodeLenses: boolean
     editorTitleCommandIcon: boolean
     experimentalGuardrails: boolean
     experimentalNonStop: boolean