+++ vscode/src/commands/menus/index.ts
@@ -129,7 +129,7 @@ export async function showCommandMenu(
 
 export async function showCustomCommandMenu(items: QuickPickItem[]): Promise<QuickPickItem> {
     const CustomCommandsMenuOptions: QuickPickOptions = {
-        title: 'Cody: Custom Commands (Experimental)',
+        title: 'Cody: Custom Commands (Beta)',
         placeHolder: 'Search command to run...',
     }
 
@@ -162,7 +162,7 @@ export async function showCustomCommandMenu(items: QuickPickItem[]): Promise<Qui
  */
 export async function showCommandConfigMenu(): Promise<CustomCommandsItem> {
     const CustomCommandConfigMenuOptions = {
-        title: 'Cody: Configure Custom Commands (Experimental)',
+        title: 'Cody: Configure Custom Commands (Beta)',
         placeHolder: 'Choose an option',
     }