+++ vscode/src/configuration.test.ts
@@ -22,7 +22,7 @@ describe('getConfiguration', () => {
                 '*': true,
                 scminput: false,
             },
-            experimentalCommandLenses: false,
+            commandCodeLenses: false,
             editorTitleCommandIcon: true,
             experimentalChatPanel: false,
             experimentalChatPredictions: false,
@@ -76,7 +76,7 @@ describe('getConfiguration', () => {
                         return true
                     case 'cody.experimental.newSearch':
                         return true
-                    case 'cody.experimental.commandLenses':
+                    case 'cody.commandCodeLenses':
                         return true
                     case 'cody.editorTitleCommandIcon':
                         return true
@@ -147,7 +147,7 @@ describe('getConfiguration', () => {
             experimentalChatPanel: true,
             experimentalChatPredictions: true,
             experimentalSearchPanel: true,
-            experimentalCommandLenses: true,
+            commandCodeLenses: true,
             editorTitleCommandIcon: true,
             experimentalGuardrails: true,
             experimentalLocalSymbols: true,